+++ BASE/Test/Microsoft.ApplicationInsights.Test/Microsoft.ApplicationInsights.Tests/Channel/TransmissionTest.cs
@@ -367,28 +367,63 @@
 
             }
 
+#if NETCOREAPP2_1
             [TestMethod]
-            public async Task SendAsyncLogsBreezeReponseTimeAndStatusCode()
+            public async Task SendAsyncLogsIngestionReponseTimeEventCounter()
             {
                 var handler = new HandlerForFakeHttpClient
                 {
                     InnerHandler = new HttpClientHandler(),
                     OnSendAsync = (req, cancellationToken) =>
                     {
-                        // VALIDATE
-                        Assert.AreEqual(testUri, req.RequestUri);
-                        Assert.AreEqual(HttpMethod.Post, req.Method);
                         return Task.FromResult<HttpResponseMessage>(new HttpResponseMessage());
                     }
                 };
 
                 using (var fakeHttpClient = new HttpClient(handler))
                 {
-                    var items = new List<ITelemetry> { new EventTelemetry(), new EventTelemetry() };
-
                     // Instantiate Transmission with the mock HttpClient
                     Transmission transmission = new Transmission(testUri, new byte[] { 1, 2, 3, 4, 5 }, fakeHttpClient, string.Empty, string.Empty);
-                    // transmission.Timeout = TimeSpan.FromMilliseconds(1);
+
+                    using (var listener = new TestEventListener())
+                    {
+                        var eventCounterArguments = new Dictionary<string, string>
+                        {
+                            {"EventCounterIntervalSec", "1"}
+                        };
+
+                        listener.EnableEvents(CoreEventSource.Log, EventLevel.LogAlways, (EventKeywords)AllKeywords, eventCounterArguments);
+
+                        HttpWebResponseWrapper result = await transmission.SendAsync();
+
+                        // VERIFY
+                        // We validate by checking SDK traces.
+                        var allTraces = listener.Messages.ToList();
+                        var traces = allTraces.Where(item => item.EventName == "EventCounters").ToList();
+                        Assert.AreEqual(1, traces.Count);
+                        var payload = (IDictionary<string, object>)traces[0].Payload[0];
+                        Assert.AreEqual("IngestionEndpoint-ResponseTimeMsec", payload["Name"].ToString());
+                        Assert.AreNotEqual(0, (float)payload["IntervalSec"]);
+                    }
+                }
+            }
+#endif
+            [TestMethod]
+            public async Task SendAsyncLogsIngestionReponseTimeAndStatusCode()
+            {
+                var handler = new HandlerForFakeHttpClient
+                {
+                    InnerHandler = new HttpClientHandler(),
+                    OnSendAsync = (req, cancellationToken) =>
+                    {
+                        return Task.FromResult<HttpResponseMessage>(new HttpResponseMessage());
+                    }
+                };
+
+                using (var fakeHttpClient = new HttpClient(handler))
+                {
+                    // Instantiate Transmission with the mock HttpClient
+                    Transmission transmission = new Transmission(testUri, new byte[] { 1, 2, 3, 4, 5 }, fakeHttpClient, string.Empty, string.Empty);
 
                     using (var listener = new TestEventListener())
                     {